Add generic code renderer for all installed languages
